3_CloudSQL_Database/requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
pyarrow
fsspec
gcsfs
psycopg2-binary
sqlalchemy
google-cloud-secret-manager